+++ src/templates/MeetingCreate/MeetingEditTemplate.tsx
@@ -1,22 +1,17 @@
-import { Button, Modal, TextField, Typography } from '@mui/material';
+import { Button, TextField, Typography } from '@mui/material';
 import dayjs, { Dayjs } from 'dayjs';
 import { useMemo, useState } from 'react';
 import { SetterOrUpdater } from 'recoil';
 
 import { DateInput } from '../../components/DateInput';
-import { MaskingInput } from '../../components/MaskingInput';
 import { Contents, Footer, Header, HeaderContainer } from '../../components/pageLayout';
 import { IMeetingEditStep } from '../../hooks/useMeetingEdit';
 import { CreateMeetingState } from '../../stores/createMeeting';
+import { InputPasswordModal } from './InputPasswordModal';
 import { MeetingEditStepper } from './MeetingEditStepper';
 import { SelectDates } from './SelectDates';
 import { SelectMeetingType } from './SelectMeetingType';
-import {
-  BorderLinearProgress,
-  FullHeightButtonGroup,
-  PasswordInput,
-  PasswordInputContainer,
-} from './styled';
+import { BorderLinearProgress, FullHeightButtonGroup } from './styled';
 
 export interface ICreateMeetingTemplateProps {
   currentStep: number;
@@ -54,14 +49,7 @@ export function MeetingEditTemplate({
   };
 
   const [showMaskingInput, setShowMaskingInput] = useState(false);
-  const [passwordText, setPasswordText] = useState('');
 
-  const onEndCreate = () => {
-    /**
-     * @TODO
-     * API 요청, 응답 시 리다이렉팅
-     */
-  };
   return (
     <>
       <Header>
@@ -99,30 +87,7 @@ export function MeetingEditTemplate({
           )}
         </FullHeightButtonGroup>
       </Footer>
-      <Modal open={showMaskingInput}>
-        <PasswordInputContainer>
-          <PasswordInput>
-            <Typography variant="h6" fontWeight={300}>
-              비밀번호를 설정할 수 있어요.
-            </Typography>
-            <Typography variant="subtitle1" fontWeight={300}>
-              모임을 수정하거나 확정할 떄 사용해요
-            </Typography>
-            <MaskingInput length={4} text={passwordText} setText={setPasswordText} />
-            <div style={{ height: 46 }}>
-              <FullHeightButtonGroup
-                fullWidth
-                disableElevation
-                variant="contained"
-                aria-label="Disabled elevation buttons"
-              >
-                <Button onClick={onEndCreate}>생략하기</Button>
-                <Button onClick={onEndCreate}>설정하기</Button>
-              </FullHeightButtonGroup>
-            </div>
-          </PasswordInput>
-        </PasswordInputContainer>
-      </Modal>
+      <InputPasswordModal showMaskingInput={showMaskingInput} />
     </>
   );
 }
